\name{MR.MCEM-package}
\alias{MR.MCEM-package}
\alias{MR.MCEM}
\docType{package}
\title{
  MC-EM Algorithm for Bayesian Mendelian Randomization
}
\description{
  MC-EM algorithm for Bayesian model that accounts for effect heterogeneity in two-sample summary-data Mendelian Randomization analysis
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Daniel Iong, daniong@umich.edu

Maintainer: Daniel Iong <daniong@umich.edu>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
