\name{MRPATH-package}
\alias{MRPATH-package}
\alias{MRPATH}
\docType{package}
\title{
  MR-PATH: A Latent Mixture for Heterogenous Causal Mechanisms in Mendelian Randomization
}
\description{
  Mendelian Randomization (MR) is a popular method in epidemiology and genetics that uses genetic variation as instrumental variables for causal inference. This package implements methods for fitting the MR-PATH model from \insertCite{iong2020latent}{MRPATH}.
}
% \details{
%   This section should provide a more detailed overview of how to use the
%   package, including the most important functions.
%
%   \insertNoCite{iong2020latent}
% }
\author{
Daniel Iong, daniong@umich.edu

Qingyuan Zhao, qyzhao@statslab.cam.ac.uk

Yang Chen, ychenang@umich.edu

Maintainer: Daniel Iong <daniong@umich.edu>
}
\references{
  \insertAllCited{}
}
\keyword{ package }
% \seealso{
% 
% }
% \examples{
%   \dontrun{
%      ## Optional simple examples of the most important functions
%      ## These can be in \dontrun{} and \donttest{} blocks.
%   }
% }
